Split inference logic into modular files
